### 2020-08-09 #2559 @diehlpk
> ```
> Reference check summary:
>
> OK DOIs
>
> - 10.1080/10618600.2016.1172487 is OK
> - 10.1007/978-1-4684-9393-1 is OK
> - 10.1111/j.1467-9868.2009.00736.x is OK
> - 10.1063/1.1699114 is OK
> - 10.1111/1467-9868.00363 is OK
> - 10.1214/aoms/1177699147 is OK
> - 10.1109/5.18626 is OK
> - 10.1049/ip-f-2.1993.0015 is OK
> - 10.1109/TSP.2005.849185 is OK
> - 10.3150/14-BEJ666 is OK
> - 10.2139/ssrn.2386371 is OK
>
> MISSING DOIs
>
> - https://doi.org/10.1093/biomet/57.1.97 may be missing for title: Monte Carlo Sampling Methods Using Markov Chains and Their Applications
> - https://doi.org/10.1109/9780470544334.ch9 may be missing for title: A New Approach to Linear Filtering and Prediction Problems
> - https://doi.org/10.1080/01621459.1999.10474153 may be missing for title: Filtering via Simulation: Auxiliary Particle Filters
>
> INVALID DOIs
>
> - None
> ```

@tbrown122387 Can you please add the missing DOIs to the paper?
