+++ paper/paper.bib
@@ -158,6 +158,7 @@ archivePrefix = "arXiv",
  pages = {97--109},
  publisher = {[Oxford University Press, Biometrika Trust]},
  title = {Monte Carlo Sampling Methods Using Markov Chains and Their Applications},
+ doi = {10.1093/biomet/57.1.97},
  volume = {57},
  year = {1970}
 }
@@ -227,6 +228,7 @@ year = {2002},
     Volume = {82},
     Number = {Series D},
     Pages = {35--45},
+    doi = {10.1109/9780470544334.ch9},
     Year = {1960}
 }
 
@@ -315,6 +317,7 @@ month={Feb},}
   bibsource =    "http://www.amstat.org/publications/jasa/;
                  http://www.math.utah.edu/pub/tex/bib/jamstatassoc.bib",
   acknowledgement = ack-nhfb,
+  doi = {10.1080/01621459.1999.10474153},
   fjournal =     "Journal of the American Statistical Association",
   journal-URL =  "http://www.tandfonline.com/loi/uasa20",
 }
