### 2020-08-10 #2559 @diehlpk
Hi @wrathematics @betatim @jochym @dvalters @jwuttke @leios would you be interested in reviewing this paper for JOSS?
