### 2020-08-10 #2559 @teuben
Not reviewing, but I'm offering one comment. I have no clue from the README.md first paragraph what this is about. Is a particle an object at some position with lots of attributes and you filter one that? If so ,say that.
