### 2020-08-11 #2559 @tbrown122387
@diehlpk To include the definition of a particle? Or the definition of “particle filter?” Regarding the first one, @teuben has it right—they are just samples targeting a distribution, and the distribution could indeed describe unobserved “positions.”
