### 2020-08-11 #2559 @diehlpk
@tbrown122387 I think it would be good to add the definition of particle, since different scientific fields have different meanings for particles. So I think it would be beneficial to make clear what you define as particles. I think also adding the term particle filter would help, because here different scientific fields have different interpreataions.
