### 2020-08-11 #2559 @teuben
If it's that general a filter, why isn't this just a table?  The "particle" is then a row in the table, it has many attributes, and you can filter them.   Is that description as accurate?  I guess I'm asking, what make your row a particle, and not a row in a table. Are there other properties linking the rows?
