### 2020-08-12 593afcb
found typos in paper, and remove unnecessary build folder
