+++ build/test/CMakeFiles/pf_test.dir/cmake_clean.cmake
@@ -1,13 +1,0 @@
-file(REMOVE_RECURSE
-  "CMakeFiles/pf_test.dir/test-main.cpp.o"
-  "CMakeFiles/pf_test.dir/test_resamplers.cpp.o"
-  "CMakeFiles/pf_test.dir/test_rv_eval.cpp.o"
-  "CMakeFiles/pf_test.dir/test_rv_samp.cpp.o"
-  "pf_test"
-  "pf_test.pdb"
-)
-
-# Per-language clean rules from dependency scanning.
-foreach(lang CXX)
-  include(CMakeFiles/pf_test.dir/cmake_clean_${lang}.cmake OPTIONAL)
-endforeach()

[…]
