### 2020-08-12 #2559 @tbrown122387
If I understand you correctly @teuben it isn't really a table. That analogy wouldn't work if each sample had dimension greater than 1, but more importantly, you're not really adding "rows."

The primary bits of data in each class are two "rows" (`std::array`s). One stores samples that target the filtering distribution (I also suspect we're not using the same definition of that word, too), and the other stores weights for those samples. All of the samples and the weights change at each time point too, after you call `filter`, and update them based on a new piece of information.

These particles (or weighted-samples) are used to approximate the filtering distribution. Time t's filtering distribution is the distribution of the log-volatility at time t, given all of the observed information up until that time point: p(x_t | y_1, ... y_t) t moves forward, and you get to approximate the sequence of these distributions in real-time. The memory requirements don't grow, and neither does the dimension of the distribution you're targeting. How and when they fail is more of a statistical question, that's definitely outside the scope of this paper.

In the particular example in the paper, the samples would represent a given time t's  log-volatility (x_t). That's model-specific, though. Its meaning depends on the model. Without using any financial jargon, this is important because you would get up-to-date estimates of how "risky" the stock market is.

Does this help?
