### 2020-08-12 #2559 @diehlpk
Hi @jordigh @hausen @pragyansmita @conradsnicta @dvalters @abhishekbajpayee @jmbr @Volkerschmid  would you be interested in reviewing this paper for JOSS?
