### 2020-08-12 #2559 @teuben
@tbrown122387 I think my background is playing havoc with me, since I do (astrophysical) N-body simulations.  Reading on one particular wikipedia page particle was defined as a candidate solution, and I could see that if you have a swarm of solutions and want to find the optimal one, filtering is one way to weed out the bad cases. So if it's the nomenclature, I can begin to understand your lingo. If so, would that help to add this definition near the top of what people read when encountering this code?
