### 2020-08-12 #2559 @jmbr
@diehlpk I can't this time around but I remain interested in reviewing for JOSS in the future.
