### 2020-08-12 #2559 @tbrown122387
@teuben yes it would. I myself have a hard time understanding any of the search results obtained from my searching "N-body simulations." I just linked in the Wiki page in the first sentence.
