### 2020-08-17 #2559 @diehlpk
Hi @pboesu @ziotom78 @zhampel @nespinoza @williamjameshandley @adavidzh @andremrsantos @fpl  would you be interested in reviewing this paper for JOSS?
