### 2020-08-18 #2559 @adavidzh
> Hi @pboesu @ziotom78 @zhampel @nespinoza @williamjameshandley @adavidzh @andremrsantos @fpl would you be interested in reviewing this paper for JOSS?

I'm afraid I have to sit this one out, @diehlpk. Thank you for asking.
