### 2020-08-20 #2559 @tbrown122387
I searched "statist*" [here](https://joss.theoj.org/about#editorial_board) and I noticed several editors who have experience with statistical techniques: fboehm, mikldk, mjsottile, marcosvital, usethedata @diehlpk you are the only editor who mentions c++ explicitly in his/her bio on that page, although many editors have github repos whose primary language is c++
