### 2020-08-20 #2559 @diehlpk
@tbrown122387 I asked for people of your community to review this paper? The editors do normally do not have the capactiy to review papers and edit them.
