### 2020-08-20 #2559 @tbrown122387
@diehlpk understood. There are no members of my professional network that are familiar with JOSS’ requirements, so I have no suggestions.
