### 2020-08-24 #2559 @ziotom78
> Hi @pboesu @ziotom78 @zhampel @nespinoza @williamjameshandley @adavidzh @andremrsantos @fpl would you be interested in reviewing this paper for JOSS?

Hi @diehlpk , sorry for the delay, but I'm currently on vacation till mid-September. I could work on this only once I came back to work, don't know if this fits the time scale of this review.
