### 2020-08-24 #2559 @andremrsantos
> Hi @pboesu @ziotom78 @zhampel @nespinoza @williamjameshandley @adavidzh @andremrsantos @fpl would you be interested in reviewing this paper for JOSS?

Sorry for the delay. I would be glad to review the paper if you're still accepting reviewers. If so, could you please link me to the journal review criteria and guidelines?
