### 2020-08-24 #2559 @diehlpk
@andremrsantos

Please find the review criteria here

https://joss.readthedocs.io/en/latest/review_criteria.html

and the review checklist here

https://joss.readthedocs.io/en/latest/review_checklist.html
