### 2020-08-24 #2599 @whedon
[REVIEW]: A Short Introduction to PF: A C++ Library for Particle Filtering
**Submitting author:** @tbrown122387 (<a href="http://orcid.org/0000-0003-4972-6251">Taylor Brown</a>)
**Repository:** <a href="https://github.com/tbrown122387/pf" target ="_blank">https://github.com/tbrown122387/pf</a>
**Version:** v1.0.3
