### 2020-09-09 #2599 @andremrsantos
Dear all,
here are some comments regarding the manuscript and the software package.

Overall, the PF library developed by the authors is a useful and interesting package for development of Particle Filters. I believe the code and document are suitable for publication after answer some of the comments below regarding the manuscript and code base.

## Comments

### Documentation

- [ ] **Installation instructions**: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.
  - On the project github page, the authors should make clear any requirements. Currently, the authors mentions the library requires _Boost_ , _Eigen_, and _Catch2_, but they appear to be only necessary for the second installation option. Also versions required are not clear.
- [ ] **Example usage**: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).
  - The authors provide an example usage, but they should include the expected results so the users can check the code behave as expected. Also, they should define a seed on their examples to make the results replicable.
- [ ] **Automated tests**: Are there automated tests or manual steps described so that the functionality of the software can be verified?
  - Automated test are included in the package, but I was unable to execute them with `make test` on the build directory. The authors should include a quick guide to run their tests.
- [ ] **Community guidelines**: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
  - I couldn't find any guidelines for third parties wishing to contribute, report issues or seek support.

### Software paper

  - The authors could expand a little on the Particle Filtering subject.
- [ ] **State of the field**: Do the authors describe how this software compares to other commonly-used packages?
  - Although the authors cite some other similar software, no other context is given except for those tools aimed to be used with scripting languages.
  - The authors also claims the library attempts to provide speed and abstraction to mitigate the complexity of particle filters and slow, but the manuscript doesn't include any comparison in terms of speed to similar projects.
- [ ]  Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?
  - Some minor issues:
  - (i) on the example section (at page 2), the example model lists phi = .91, but on the text says it is .95;
  - (ii) I not particular sure why an object-oriented design facilitates implementation of update to many particular in place of others design models. I would suggested the authors to justify or remove the statement.

Best,
André Santos
