### 2020-09-12 b4bd275
add test-running instructions to README
