### 2020-09-12 712814c
add dependency version numbers
