### 2020-09-12 tbrown122387/pf#11 @tbrown122387
add seed-setting capabilities in resampler base class
Users will want to be able to set seeds to check reproducibility. That should be added as a capability in the `rbase` base class template.
