### 2020-09-12 6a138cc
add particle filtering details to paper
