### 2020-09-12 tbrown122387/pf#12 @tbrown122387
add speed comparison example
add speed comparison example to demonstrate that it's faster than other libraries
