### 2020-09-12 #2599 @tbrown122387
@andremrsantos thanks very much for your work reviewing my paper. I found the feedback spot-on and very useful.

For the sake of organization, I'll respond to your suggestions point by point.

First, there were some changes I was able to make relatively quickly:

> On the project github page, the authors should make clear any requirements. Currently, the authors mentions the library requires Boost , Eigen, and Catch2, but they appear to be only necessary for the second installation option. Also versions required are not clear.

I added a section to the README that precedes the installation instructions and that mentions dependencies with specific versions: [here](https://github.com/tbrown122387/pf/commit/4885f5fdf053cf62a2226e813dff7a6ea190cad1)

I also added specific version numbers to the installation section: [see here](https://github.com/tbrown122387/pf/commit/712814cc326d5034866e44141a20fd08422616f1)

> Automated test are included in the package, but I was unable to execute them with make test on the build directory. The authors should include a quick guide to run their tests.

Test building is done automatically with the `cmake` installation approach, but the executable is sort of hidden and not really mentioned, so you're right about that. I added a mention of where the executable would show up [here](https://github.com/tbrown122387/pf/commit/b4bd27586db9970f5df47121242c7f9c5c69b70c)

> The authors could expand a little on the Particle Filtering subject.

- I added some more [details](https://github.com/tbrown122387/pf/commit/6a138cc4976fc5019dae8d605e54d95659b8496b) to the paper about what filtering and particle filtering are, and how they are useful. I can always add more.

> on the example section (at page 2), the example model lists phi = .91, but on the text says it is .95

- Nice catch. Just made the change. It is indeed supposed to be `.91`. The change is [here](https://github.com/tbrown122387/pf/commit/01de26ecec3286ba6afaf60606707f46ad49bc74)

> I not particular sure why an object-oriented design facilitates implementation of update to many particular in place of others design models. I would suggested the authors to justify or remove the statement.

I add more details about why I think this is important [here](https://github.com/tbrown122387/pf/commit/f4ebfa839e53af2ee2bc43d10ca6afb8cd145e6d) and [here](https://github.com/tbrown122387/pf/commit/01de26ecec3286ba6afaf60606707f46ad49bc74) I think it makes using particle filters inside more complicated algorithms a bit easier.

Second, these changes might take me some more time, but I've started working on them:

> The authors also claims the library attempts to provide speed and abstraction to mitigate the complexity of particle filters and slow, but the manuscript doesn't include any comparison in terms of speed to similar projects.

I'm working on a side-by-side comparison now. I've added an issue [here](https://github.com/tbrown122387/pf/issues/12)

> The authors provide an example usage, but they should include the expected results so the users can check the code behave as expected. Also, they should define a seed on their examples to make the results replicable.

I've added an issue for this as well ([here](https://github.com/tbrown122387/pf/issues/11)). So far, seeds have been set by the clock. I'll just need to add another constructor to the base class template for the resampling types.

> I couldn't find any guidelines for third parties wishing to contribute, report issues or seek support.

I'm still thinking about how to do this. I've never been able to recruit many people in real life to help me with this software, so this point is of great interest to me. I've seen other software projects with online forums...maybe that would be something worth exploring. I'm open to any suggestions.
