### 2020-09-17 #2599 @diehlpk
@tbrown122387 instead of having a shell script to run the tests, you could add ctests so the user could type make test to run them.
