### 2020-09-18 tbrown122387/pf#13 @ziotom78
Easier way to interface with catch2?
The instructions in the README ask the user to install a few libraries (boost, eigen3, catch2). However, the requirement on catch2 is not explained well:

1. Catch2's homepage explains that the package can be installed by just copying only the header file;
2. People running any Ubuntu-based distribution will likely install it using `sudo apt install catch`;
3. OTOH, the `CMakeLists.txt` provided by this package expects a few `*.cmake` files from the Catch2 distribution, which means that it requires Catch2 to be installed using a non-standard procedure. (No, `apt install` does not install these files where `cmake` can find them`.)

I see a few possible solutions:

1. Include a copy of the Catch2 header file in the source (possible if Catch2's license is compatible with this one: IANAL, so I cannot give advice here). This would be the easiest and cleanest solution;

2. Enable `cmake` to detect the presence of the Catch2 header file when the `*.cmake` files are not found

3. Explain in the README the expected way to install Catch2.

(This issue was raised while carrying on the review for JOSS: https://github.com/openjournals/joss-reviews/issues/2599.)
