### 2020-09-18 tbrown122387/pf#13 @tbrown122387
@ziotom78 regarding (3), which `*.cmake` files are you referring to? Are you referring to [this](https://github.com/tbrown122387/pf/blob/master/test/CMakeLists.txt) `CMakeLists.txt` file or [this one?](https://github.com/tbrown122387/pf/blob/master/CMakeLists.txt)
