### 2020-09-19 d491ee3
division by zero error fix
