+++ include/pf/bootstrap_filter_with_covariates.h
@@ -224,7 +224,7 @@ void BSFilterWC<nparts, dimx, dimy, dimcov, resamp_t, float_t>::filter(const osv
        
         // try to iterate over particles all at once
         ssv newSamp;
-        float_t maxOldLogUnNormWts(-1.0/0.0);
+        float_t maxOldLogUnNormWts(-std::numeric_limits<float_t>::infinity());
         arrayfloat_t oldLogUnNormWts = m_logUnNormWeights;
         for(size_t ii = 0; ii < nparts; ++ii)
         {
