### 2020-09-19 239f208
tests, dox, more division by zero errors
