+++ docs/latex/classrvsamp_1_1UnivInvGammaSampler.tex
@@ -1,0 +1,104 @@
+\hypertarget{classrvsamp_1_1UnivInvGammaSampler}{}\section{rvsamp\+:\+:Univ\+Inv\+Gamma\+Sampler$<$ float\+\_\+t $>$ Class Template Reference}
+\label{classrvsamp_1_1UnivInvGammaSampler}\index{rvsamp\+::\+Univ\+Inv\+Gamma\+Sampler$<$ float\+\_\+t $>$@{rvsamp\+::\+Univ\+Inv\+Gamma\+Sampler$<$ float\+\_\+t $>$}}
+
+
+A class that performs sampling from a univariate Inverse Gamma distribution.  
+
+
+
+
+{\ttfamily \#include $<$rv\+\_\+samp.\+h$>$}
+
+
+
+Inheritance diagram for rvsamp\+:\+:Univ\+Inv\+Gamma\+Sampler$<$ float\+\_\+t $>$\+:
+\nopagebreak
+\begin{figure}[H]
+\begin{center}
+\leavevmode
+\includegraphics[width=239pt]{classrvsamp_1_1UnivInvGammaSampler__inherit__graph}
+\end{center}
+\end{figure}
+
+
+Collaboration diagram for rvsamp\+:\+:Univ\+Inv\+Gamma\+Sampler$<$ float\+\_\+t $>$\+:
+\nopagebreak
+\begin{figure}[H]
+\begin{center}
+\leavevmode
+\includegraphics[width=239pt]{classrvsamp_1_1UnivInvGammaSampler__coll__graph}
+\end{center}
+\end{figure}
+\subsection*{Public Member Functions}
+\begin{DoxyCompactItemize}
+\item 
+\mbox{\Hypertarget{classrvsamp_1_1UnivInvGammaSampler_abb7dbe2cde9fd88880a86765b609d5ac}\label{classrvsamp_1_1UnivInvGammaSampler_abb7dbe2cde9fd88880a86765b609d5ac}} 
+\hyperlink{classrvsamp_1_1UnivInvGammaSampler_abb7dbe2cde9fd88880a86765b609d5ac}{Univ\+Inv\+Gamma\+Sampler} ()
+\begin{DoxyCompactList}\small\item\em Default-\/constructor ... \end{DoxyCompactList}\item 
+\hyperlink{classrvsamp_1_1UnivInvGammaSampler_a90cc04cdd50c15b4795b4f70dbb9dcb5}{Univ\+Inv\+Gamma\+Sampler} (float\+\_\+t alpha, float\+\_\+t beta)
+\item 
+float\+\_\+t \hyperlink{classrvsamp_1_1UnivInvGammaSampler_a5d477f520d0a5069c5fcda5a1952cead}{sample} ()
+\begin{DoxyCompactList}\small\item\em draws a random number. \end{DoxyCompactList}\end{DoxyCompactItemize}
+\subsection*{Private Attributes}
+\begin{DoxyCompactItemize}
+\item 
+\mbox{\Hypertarget{classrvsamp_1_1UnivInvGammaSampler_ab0ff81a383579c8a6e7f56355270dffe}\label{classrvsamp_1_1UnivInvGammaSampler_ab0ff81a383579c8a6e7f56355270dffe}} 
+std\+::gamma\+\_\+distribution$<$ float\+\_\+t $>$ \hyperlink{classrvsamp_1_1UnivInvGammaSampler_ab0ff81a383579c8a6e7f56355270dffe}{m\+\_\+gamma\+\_\+gen}
+\begin{DoxyCompactList}\small\item\em makes gamma random variates that we take the reciprocal of \end{DoxyCompactList}\item 
[…]
