+++ docs/latex/classkalman.tex
@@ -346,7 +346,7 @@ void \hyperlink{classkalman}{kalman}$<$ dimstate, dimobs, diminput, float\+\_\+t
 Predicts the next state. 
 
 \begin{DoxyRefDesc}{Todo}
-\item[\hyperlink{todo__todo000002}{Todo}]handle diagonal variance matrices, and ensure symmetricness in other ways \end{DoxyRefDesc}
+\item[\hyperlink{todo__todo000001}{Todo}]handle diagonal variance matrices, and ensure symmetricness in other ways \end{DoxyRefDesc}
 
 \begin{DoxyParams}{Parameters}
 {\em state\+Trans\+Mat} & \\
@@ -362,4 +362,4 @@ Predicts the next state.
 
 The documentation for this class was generated from the following file\+:\begin{DoxyCompactItemize}
 \item 
-include/\hyperlink{cf__filters_8h}{cf\+\_\+filters.\+h}\end{DoxyCompactItemize}
+include/pf/\hyperlink{cf__filters_8h}{cf\+\_\+filters.\+h}\end{DoxyCompactItemize}

[…]
