### 2020-09-20 dda32aa
add resampler constructors that accept seed
