+++ docs/latex/classBSFilter.tex
@@ -11,8 +11,7 @@ A base class for the bootstrap particle filter.
 
 
 
-Inheritance diagram for B\+S\+Filter$<$ nparts, dimx, dimy, resamp\+\_\+t, float\+\_\+t, debug $>$\+:
-\nopagebreak
+Inheritance diagram for B\+S\+Filter$<$ nparts, dimx, dimy, resamp\+\_\+t, float\+\_\+t, debug $>$\+:\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
@@ -21,8 +20,7 @@ Inheritance diagram for B\+S\+Filter$<$ nparts, dimx, dimy, resamp\+\_\+t, float
 \end{figure}
 
 
-Collaboration diagram for B\+S\+Filter$<$ nparts, dimx, dimy, resamp\+\_\+t, float\+\_\+t, debug $>$\+:
-\nopagebreak
+Collaboration diagram for B\+S\+Filter$<$ nparts, dimx, dimy, resamp\+\_\+t, float\+\_\+t, debug $>$\+:\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
