+++ docs/latex/classmn__resampler.tex
@@ -38,7 +38,9 @@ using \hyperlink{classmn__resampler_afb5d000e2464afef813792c57c42599b}{array\+In
 \item 
 \mbox{\Hypertarget{classmn__resampler_a016a00570c30806a0fdad25385395f95}\label{classmn__resampler_a016a00570c30806a0fdad25385395f95}} 
 \hyperlink{classmn__resampler_a016a00570c30806a0fdad25385395f95}{mn\+\_\+resampler} ()=default
-\begin{DoxyCompactList}\small\item\em Default constructor. Only option available. \end{DoxyCompactList}\item 
+\begin{DoxyCompactList}\small\item\em Default constructor. \end{DoxyCompactList}\item 
+\hyperlink{classmn__resampler_a179cbea35530cf7e171b1d214e1d9776}{mn\+\_\+resampler} (unsigned long seed)
+\begin{DoxyCompactList}\small\item\em Constructor that sets the seed. \end{DoxyCompactList}\item 
 void \hyperlink{classmn__resampler_a13b1897e180a791a3a099d5d6329a125}{resamp\+Log\+Wts} (\hyperlink{classrbase_aa12fc826befa6ba0647b5f59ebc396ee}{array\+Vec} \&old\+Parts, \hyperlink{classrbase_a6f76bef853e508cb5b6f546d231b06f5}{array\+Float} \&old\+Log\+Un\+Norm\+Wts)
 \begin{DoxyCompactList}\small\item\em resamples particles. \end{DoxyCompactList}\end{DoxyCompactItemize}
 \subsection*{Additional Inherited Members}
@@ -86,6 +88,25 @@ type alias for array of Eigen Matrices \mbox{\Hypertarget{classmn__resampler_a1c
 using \hyperlink{classmn__resampler}{mn\+\_\+resampler}$<$ nparts, dimx, float\+\_\+t $>$\+::\hyperlink{classrbase_ae20e0b8df15aa109252f57ecbf1f20f8}{ssv} =  Eigen\+::\+Matrix$<$float\+\_\+t,dimx,1$>$}
 
 type alias for linear algebra stuff 
+
+\subsection{Constructor \& Destructor Documentation}
+\mbox{\Hypertarget{classmn__resampler_a179cbea35530cf7e171b1d214e1d9776}\label{classmn__resampler_a179cbea35530cf7e171b1d214e1d9776}} 
+\index{mn\+\_\+resampler@{mn\+\_\+resampler}!mn\+\_\+resampler@{mn\+\_\+resampler}}
+\index{mn\+\_\+resampler@{mn\+\_\+resampler}!mn\+\_\+resampler@{mn\+\_\+resampler}}
+\subsubsection{\texorpdfstring{mn\+\_\+resampler()}{mn\_resampler()}}
+{\footnotesize\ttfamily template$<$size\+\_\+t nparts, size\+\_\+t dimx, typename float\+\_\+t $>$ \\
+\hyperlink{classmn__resampler}{mn\+\_\+resampler}$<$ nparts, dimx, float\+\_\+t $>$\+::\hyperlink{classmn__resampler}{mn\+\_\+resampler} (\begin{DoxyParamCaption}\item[{unsigned long}]{seed }\end{DoxyParamCaption})}
+
+
+
+Constructor that sets the seed. 
+
+
+\begin{DoxyParams}{Parameters}
+{\em seed} & \\
+\hline
+\end{DoxyParams}
+
 
 \subsection{Member Function Documentation}
 \mbox{\Hypertarget{classmn__resampler_a13b1897e180a791a3a099d5d6329a125}\label{classmn__resampler_a13b1897e180a791a3a099d5d6329a125}} 
