+++ docs/latex/rv__samp_8h.tex
@@ -8,16 +8,14 @@ all rv samplers must inherit from this.
 {\ttfamily \#include $<$chrono$>$}\newline
 {\ttfamily \#include $<$Eigen/\+Dense$>$}\newline
 {\ttfamily \#include $<$random$>$}\newline
-Include dependency graph for rv\+\_\+samp.\+h\+:
-\nopagebreak
+Include dependency graph for rv\+\_\+samp.\+h\+:\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
 \includegraphics[width=285pt]{rv__samp_8h__incl}
 \end{center}
 \end{figure}
-This graph shows which files directly or indirectly include this file\+:
-\nopagebreak
+This graph shows which files directly or indirectly include this file\+:\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
