+++ docs/latex/auxiliary__pf_8h.tex
@@ -12,8 +12,7 @@ A base class for Auxiliary Particle Filtering. Inherit from this if you want to 
 {\ttfamily \#include $<$cmath$>$}\newline
 {\ttfamily \#include \char`\"{}pf\+\_\+base.\+h\char`\"{}}\newline
 {\ttfamily \#include \char`\"{}rv\+\_\+samp.\+h\char`\"{}}\newline
-Include dependency graph for auxiliary\+\_\+pf.\+h\+:
-\nopagebreak
+Include dependency graph for auxiliary\+\_\+pf.\+h\+:\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode

[…]
