### 2020-09-21 tbrown122387/pf#13 @ziotom78
@tbrown122387 , I was quoting the error message produced by `cmake`:

```
$ cmake ..
CMake Error at CMakeLists.txt:11 (find_package):
  By not providing "FindCatch2.cmake" in CMAKE_MODULE_PATH this project has
  asked CMake to find a package configuration file provided by "Catch2", but
  CMake did not find one.

  Could not find a package configuration file provided by "Catch2" (requested
  version 2.9.2) with any of the following names:

    Catch2Config.cmake
    catch2-config.cmake

  Add the installation prefix of "Catch2" to CMAKE_PREFIX_PATH or set
  "Catch2_DIR" to a directory containing one of the above files.  If "Catch2"
  provides a separate development package or SDK, be sure it has been
  installed.

-- Configuring incomplete, errors occurred!
See also "/home/tomasi/Documents/work/articles/referee/JOSS/pf/pf/build/CMakeFiles/CMakeOutput.log".
```

Note that I already downloaded and manually installed `catch2/catch.hpp` (following the [instructions on their site](https://github.com/catchorg/Catch2/blob/master/docs/tutorial.md#getting-catch2)):
```
$ sudo mkdir -p /usr/include/catch2
$ curl https://raw.githubusercontent.com/catchorg/Catch2/master/single_include/catch2/catch.hpp \
  | sudo tee /usr/include/catch2/catch.hpp > /dev/null
$  head /usr/include/catch2/catch.hpp
/*
 *  Catch v2.13.1
 *  Generated: 2020-09-07 12:12:38.090364
 *  ----------------------------------------------------------
 *  This file has been merged from multiple headers. Please don't edit it directly
 *  Copyright (c) 2020 Two Blue Cubes Ltd. All rights reserved.
 *
 *  Distributed under the Boost Software License, Version 1.0. (See accompanying
 *  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 */
$
```
