### 2020-09-22 #2599 @andremrsantos
Dear @tbrown122387, here are some considerations regarding your answer:

- **Contribution Guidelines**
  - I think you should just include a simple guideline such as [Rails](https://github.com/rails/rails/blob/master/CONTRIBUTING.md) listed as example in the [page](https://docs.github.com/en/github/building-a-strong-community/setting-guidelines-for-repository-contributors) you sent, with some suggestions of how to open an issue, send a PR request, and include a basic code of conduct. Nothing much complicate.
- **I will wait for updates on the other topics you mention it may require some time**.
