### 2020-09-25 tbrown122387/pf#13 @tbrown122387
Catch2 describes two installation options [here](https://github.com/catchorg/Catch2/blob/master/docs/tutorial.md#getting-catch2). I prefer to ask users to [install the CMake package](https://github.com/catchorg/Catch2/blob/master/docs/cmake-integration.md#installing-catch2-from-git-repository). I think this will allow everything to build unchanged. @ziotom78
