+++ CONTRIBUTING.md
@@ -1,0 +1,41 @@
+## How to contribute to `pf`
+
+#### **Contribute a fully-worked example!**
+
+* send us a link to a Github repository with a self-contained example, and after we're able to build the project, we'll be sure to link to it in this project's README!
+
+#### **Did you find a bug?**
+
+* **Ensure the bug was not already reported** by searching on GitHub under [Issues](https://github.com/tbrown122387/pf/issues).
+
+* If you're unable to find an open issue addressing the problem, [open a new one](https://github.com/tbrown122387/pf/issues/new) and label it as a "bug." 
+
+* Be sure to include a **title and clear description**, as much relevant information as possible, and a **code sample** or an **executable test case** demonstrating the expected behavior that is not occurring.
+
+#### **Did you fix a bug?**
+
+* Open a new GitHub pull request with the patch.
+
+* Ensure the PR description clearly describes the problem and solution. Include the relevant issue number if applicable.
+
+* Changes that are cosmetic in nature and do not add anything substantial to the stability, functionality, or testability of Rails will generally not be accepted.
+
+#### **Do you intend to add a new feature or change an existing one?**
+
+* Suggest your change by opening up an [issue](https://github.com/tbrown122387/pf/issues), and labeling it as an "enhanement."
+
+* Open a new GitHub pull request with the enhancement.
+
+#### **Do you have questions about the source code or how to incorporate `pf` into your project?**
+
+* Ask any question you have by opening up an [issue](https://github.com/tbrown122387/pf/issues), and labeling it as a "question."
+
+#### **Do you want to contribute to the documentation?**
+
+* Awesome! Documentation is automatically generated with [Doxygen.](https://www.doxygen.nl/manual/docblocks.html), so most of that work will involve editing comments in header files.
+
+* Running `doxygen Doxyfile` will generate all files in `docs/`
+ 
+* Submit a pull request when you're finished.
+
+Thanks! 
