### 2020-09-25 7e7d0f5
update README and fix typo in CONTRIBUTING.md
