### 2020-09-25 tbrown122387/pf#13 @ziotom78
@tbrown122387 , ok, I believe it's fine, as long as this is explicitly stated in the documentation.
