### 2020-09-25 e521314
add specific catch2 installation instructions
