### 2020-09-28 #2599 @diehlpk
@tbrown122387  I checked the references and please update

    Murray, L. M. (2013). Bayesian state-space modelling on high-performance hardware using LibBi. (add the arxiv.org information)
    Taylor, S. (1982). Financial returns modelled by the product of two stochastic processes, a study of daily sugar prices 1961-79,1. (Add the journal)
    Brown, T. R. (2020). Approximating posterior predictive distributions by averaging output from many particle filters. (Add archiv.org information)
