### 2020-09-29 #2599 @tbrown122387
@ziotom78 thanks very much for your very helpful and valuable comments! @diehlpk so far I have addressed all of the problems except for two: updating the references and writing a speed comparison example. I should be done relatively soon. Apologies for the delay--this is a busy week for me.
