+++ test/test_rv_samp.cpp
@@ -19,15 +19,22 @@ public:
 //    MVNSampler<bigdim> m_mns2; // nondefault construction  TODO add
     rvsamp::UniformSampler<double> m_us;
     rvsamp::UniformSampler<double> m_us2; // nondefault construction
-    
-    SampFixture() : m_us2(-2.0, -1.0) // weirder upper and lower bounds
+    rvsamp::UnivStudTSampler<double> m_t_sampler;
+   
+ 
+  
+  
+   SampFixture() 
+        : m_us2(-2.0, -1.0) // weirder upper and lower bounds
+        , m_t_sampler(2.0)
     {
         m_ns.setMean(2.0);
         m_ns.setStdDev(1.5);
         m_mns.setMean(Vec::Zero());
         m_mns.setCovar(Mat::Identity());
     }
-    
+
+
 };
 
 
@@ -51,3 +58,12 @@ TEST_CASE_METHOD(SampFixture, "UniformTest", "[samplers]")
     REQUIRE(-2.0 < m_us2.sample());
     REQUIRE( m_us2.sample() < -1.0);
 }
+
+
+TEST_CASE_METHOD(SampFixture, "Student T Test", "[samplers]")
+{
+
+    REQUIRE( -std::numeric_limits<double>::infinity() < m_t_sampler.sample());
+}
+
+
