### 2020-10-01 e3e9e98
first draft of smctc comparison example
