### 2020-10-04 fde23e5
update references and clean comparison example
