+++ paper/paper.md
@@ -23,21 +23,21 @@ In this library, each available particle filtering algorithm is provided as an a
 
 The details of each algorithm are abstracted away, and each algorithm's class template's required functions are pure virtual methods, meaning that the data analyst will not be able to omit any function that is required by the algorithm. 
 
-This is by no means the first ``C++`` library to be offered that provides particle filter implementations. Other options include ``LibBi`` [@libbi], ``Biips`` [@todeschini2014biips], ``SMCTC`` [@Johansen:2009:JSSOBK:v30i06] and ``Nimble`` [@doi:10.1080/10618600.2016.1172487]. The goals of these software packages are different, though--users of these libraries write their models in a scripting language, and that model file gets parsed into ``C++`` code. This library is designed for users that prefer to work in ``C++`` directly.
+This is by no means the first ``C++`` library to be offered that provides particle filter implementations. Other options include ``LibBi`` [@libbi], ``Biips`` [@todeschini2014biips], ``Nimble`` [@doi:10.1080/10618600.2016.1172487] and ``SMCTC`` [@Johansen:2009:JSSOBK:v30i06]. The goals of most of these software packages are different, though--users of these libraries typically write their models in a scripting language, and that model file gets parsed into ``C++`` code. This library, on the other hand, is designed for users that prefer to work in ``C++`` directly. 
 
 # Statement of Need
 
-State-space models describe a partially-observed Markov chain--they possess a hidden/latent variable at each time point, in addition to an observed variable. ``Filtering" is defined as obtaining the distributions of each unobserved state/code random variable, conditioning on all of the observed information up to that point in time. ``Particle filters" are a class of algorithms that approximate this sequence of distributions with weighted samples (termed particles). Filtering is a useful tool for a variety of applications in a variety of fields, but it should also be mentioned that they can be used for real-time forecasting with only minor modifications, and they are critical component of more advanced parameter estimation algorithms. 
+State-space models describe a partially-observed Markov chain that possess a hidden/latent variable at each time point, in addition to an observed variable. ``Filtering" is defined as obtaining the distributions of each unobserved state/code random variable, conditioning on all of the observed information up to that point in time. ``Particle filters" are a class of algorithms that approximate this sequence of distributions with weighted samples (termed particles). Filtering is a useful tool for a variety of applications in a variety of fields, and it should also be mentioned that they can be used for real-time forecasting, and they are critical component of more advanced parameter estimation algorithms. 
 
 Unfortunately, it takes time and effort to implement different particle filters well, and this is true for two reasons. First, the mathematical notation used to describe them can be complex. The second reason is that, even if they are correctly implemented, they can be quite slow, limiting the number of tasks that they would be feasible for. This library attempts to provide speed and abstraction to mitigate these two difficulties.
 
-Additionally, this software is designed in an object-oriented manner. This allows for individual particle filters to be used in isolation, and it also facilitates the implementation of more complicated algorithms that update many particle filters through many iterations in a stateful way, possibly in parallel. For this second class of algorithms, there are usually two necessary loops: the ``outer" loop that loops over time, and the ``inner" loop that iterates over each distinct particle filter. Without an object-oriented design, there would be a third loop, which loops over all particle samples in each particle filter. Some examples of opportunities include particle filters with parallelized resampling schemes  [@1453776,1309.2918], particle Markov chain Monte Carlo algorithms [@pmcmc], importance sampling "squared" [@issquared], and the particle swarm algorithm [@pswarm]. 
[…]
