+++ paper/paper.bib
@@ -25,6 +25,7 @@ Author = {Lawrence M. Murray},
 Title = {Bayesian State-Space Modelling on High-Performance Hardware Using \pkg{LibBi}},
 Year = {2013},
 Eprint = {arXiv:1306.3277},
+journal={arXiv preprint arXiv:1306.3277},
 }
 
 
@@ -184,16 +185,19 @@ archivePrefix = "arXiv",
   year = 2012
 }
 
-@article{taylor82,
-author = {Taylor, Stephen},
-year = {1982},
-month = {01},
-pages = {},
-title = {Financial Returns Modelled by the Product of Two Stochastic Processes, a Study of Daily Sugar Prices 1961-79},
-volume = {1},
-booktitle = {Time Series Analysis: Theory and Practice}
+@inbook{taylor82,
+title = "Financial returns modelled by the product of two stochastic processes, a study of daily sugar prices",
+author = "Taylor, {Stephen John}",
+note = "Reprinted, original publication 1982, in Time Series Analysis: Theory and Practice 1",
+year = "2018",
+language = "English",
+isbn = "9781788110617",
+series = "The International Library of Critical Writings in Economics ",
+publisher = "Edward Elgar",
+pages = "423--446",
+editor = "Torben Andersen and Tim Bollerslev",
+booktitle = "Volatility",
 }
-
 
 @article {rbpfandrieudoucet,
 author = {Andrieu, Christophe and Doucet, Arnaud},
@@ -359,4 +363,5 @@ Author = {Taylor R. Brown},
 Title = {Approximating Posterior Predictive Distributions by Averaging Output From Many Particle Filters},
 Year = {2020},
 Eprint = {arXiv:2006.15396},
+journal={arXiv preprint arXiv:2006.15396},
 }
