+++ paper/paper.md
@@ -27,7 +27,7 @@ This is by no means the first ``C++`` library to be offered that provides partic
 
 # Statement of Need
 
-State-space models describe a partially-observed Markov chain that possess a hidden/latent variable at each time point, in addition to an observed variable. ``Filtering" is defined as obtaining the distributions of each unobserved state/code random variable, conditioning on all of the observed information up to that point in time. ``Particle filters" are a class of algorithms that approximate this sequence of distributions with weighted samples (termed particles). Filtering is a useful tool for a variety of applications in a variety of fields, and it should also be mentioned that they can be used for real-time forecasting, and they are critical component of more advanced parameter estimation algorithms. 
+State-space models describe a partially-observed Markov chain that possess a hidden/latent variable at each time point, in addition to an observed variable. "Filtering" is defined as obtaining the distributions of each unobserved state/code random variable, conditioning on all of the observed information up to that point in time. ``Particle filters" are a class of algorithms that approximate this sequence of distributions with weighted samples (termed particles). Filtering is a useful tool for a variety of applications in a variety of fields, and it should also be mentioned that they can be used for real-time forecasting, and they are critical component of more advanced parameter estimation algorithms. 
 
 Unfortunately, it takes time and effort to implement different particle filters well, and this is true for two reasons. First, the mathematical notation used to describe them can be complex. The second reason is that, even if they are correctly implemented, they can be quite slow, limiting the number of tasks that they would be feasible for. This library attempts to provide speed and abstraction to mitigate these two difficulties.
 
