### 2020-10-05 #2599 @diehlpk
@tbrown122387

Following remarks:

I think, that it would be beneficial for the reader to have a link to the README.md file here

`this is “Option 2” described in theREADME.mdfile`

and same for examples/svol_sisr.h here

`The fileexamples/svol_sisr.hprovides an example of writing a class template calledsvol_sisrfor this model-algorithm pair`

To make the definitions more prominent, would you like to consider to have a bullet list here

`Filtering” is defined asobtaining the distributions of each unobserved state/code random variable, conditioning onall of the observed information up to that point in time. “Particle filters” are a class ofalgorithms that approximate this sequence of distributions with weighted samples (termedparticles)`

Or any other way to highlight the definitions?
