+++ paper/paper.md
@@ -27,13 +27,17 @@ This is by no means the first ``C++`` library to be offered that provides partic
 
 # Statement of Need
 
-State-space models describe a partially-observed Markov chain that possess a hidden/latent variable at each time point, in addition to an observed variable. "Filtering" is defined as obtaining the distributions of each unobserved state/code random variable, conditioning on all of the observed information up to that point in time. ``Particle filters" are a class of algorithms that approximate this sequence of distributions with weighted samples (termed particles). Filtering is a useful tool for a variety of applications in a variety of fields, and it should also be mentioned that they can be used for real-time forecasting, and they are critical component of more advanced parameter estimation algorithms. 
+State-space models describe a partially-observed Markov chain $\{(x_t,y_t)\}_{t \ge 1}$ that possess a hidden/latent variable at each time point (denoted by $x_t$), in addition to an observed variable (denoted by $y_t$). "Filtering" is defined as obtaining the distribution of each unobserved state/code random variable, conditioning on all of the observed information up to that point in time. In other words, it is the task of finding 
+\begin{eqnarray}
+p(x_t \mid y_t, y_{t-1}, \ldots, y_1).
+\end{eqnarray}
+``Particle filters" are a class of algorithms that approximate this sequence of distributions with weighted samples (termed particles). Filtering is a useful tool for a variety of applications in a variety of fields, and it should also be mentioned that they can be used for real-time forecasting, and they are critical component of more advanced parameter estimation algorithms. 
 
 Unfortunately, it takes time and effort to implement different particle filters well, and this is true for two reasons. First, the mathematical notation used to describe them can be complex. The second reason is that, even if they are correctly implemented, they can be quite slow, limiting the number of tasks that they would be feasible for. This library attempts to provide speed and abstraction to mitigate these two difficulties.
 
 Additionally, this software is designed in an object-oriented manner. This allows for individual particle filters to be used in isolation, and it also facilitates the implementation of more complicated algorithms that update many particle filters through many iterations in a stateful way, possibly in parallel. For this second class of algorithms, there are usually two necessary loops: the "outer" loop that loops over time, and the "inner" loop that iterates over each distinct particle filter. Without an object-oriented design, there would be a third loop, which loops over all particle samples in each particle filter. Some examples of algorithms that run many particle filter within the overall algorithm are are particle filters with parallelized resampling schemes  [@1453776,1309.2918], particle Markov chain Monte Carlo algorithms [@pmcmc], importance sampling "squared" [@issquared], and the particle swarm algorithm [@pswarm]. 
 
-Finally, this library is "header-only." This will allow some users to build their own ``C++`` project by `#include`-ing relevant headers from this library and pointing the compiler at the `include/pf/` directory (this is "Option 2" described in the `README.md` file). On the other hand, other uses will prefer to use the `CMake` build procedure (this is "Option 1" in the `README.md` file), as this will help build the included unit tests and more closely follows the style of the provided example code projects. 
+Finally, this library is "header-only." This will allow some users to build their own ``C++`` project by `#include`-ing relevant headers from this library and pointing the compiler at the `include/pf/` directory (this is "Option 2" described in the [`README.md`](https://github.com/tbrown122387/pf/blob/master/README.md) file). On the other hand, other uses will prefer to use the `CMake` build procedure (this is "Option 1" in the `README.md` file), as this will help build the included unit tests and more closely follows the style of the provided example code projects. 
 
 # Examples
 
@@ -51,7 +55,7 @@ x_t = \phi x_{t-1} + \sigma z'_t.
 \end{eqnarray}
[…]
