### 2020-10-06 #2599 @tbrown122387
@diehlpk  perhaps I could add an equation/some notation? Eyes are drawn to equations, and it would also cement some of the notation that would be (re-)used in the coming Example section. It might also be easier on the reader to see me label states/codes as x_t immediately after the first time I use those words, and to see "observed information" labeled with $y_{1:t}$ immediately after they see that phrase for for the first time too. Let me know what you think of the most recent commit. I added the links as well
