### 2020-10-06 #2599 @diehlpk
@tbrown122387

`perhaps I could add an equation/some notation?`

Sounds good.
