### 2020-10-06 #2599 @diehlpk
@tbrown122387

The next steps are

- [ ] Generate a new release including all changes during the review
- [ ] Post the release here
- [ ] Generate a DOI of the release on Figshare or Zenodo. Note that the name has to match the paper title.
- [ ]  Post the DOI here
