### 2020-10-06 #2599 @tbrown122387
@diehlpk

- the latest release can be found [here](https://github.com/tbrown122387/pf/releases)
- the DOI can be found [here](https://zenodo.org/record/4068564#.X3ybEZopB2V)
- I changed the DOI's "Title" to "A Short Introduction to PF: A C++ Library for Particle Filtering", but might have done this incorrectly. Please let me know if I need to make another change.
