### 2020-10-06 #2599 @whedon
:wave: @openjournals/joss-eics, this paper is ready to be accepted and published.
