### 2020-10-07 #2599 @diehlpk
You still can update typos in the paper. However, the typos will remain in the release.
