### 2020-10-07 #2599 @kthyng
Version is up to date, and zenodo archive metadata looks good!

The paper looks good, except for the references, which have lost capitalization. @tbrown122387 can you go through your reference list in detail to preserve capitalization? You can do this by placing {} around the words in the .bib file that have capitalization that is not being preserved. {Markov} and {Monte Carlo} are what I've seen in a couple of the references, but I stopped looking so there may be more.
