### 2020-10-08 711c348
uppercase fixing in bib
