+++ paper/paper.bib
@@ -22,7 +22,7 @@
 
 @misc{libbi,
 Author = {Lawrence M. Murray},
-Title = {Bayesian State-Space Modelling on High-Performance Hardware Using \pkg{LibBi}},
+Title = {Bayesian {State-Space} {Modelling} on {High-Performance} {Hardware} {Using} \pkg{LibBi}},
 Year = {2013},
 Eprint = {arXiv:1306.3277},
 journal={arXiv preprint arXiv:1306.3277},
@@ -30,7 +30,7 @@ journal={arXiv preprint arXiv:1306.3277},
 
 
 @article{todeschini2014biips,
-  title={\pkg{Biips}: Software for {B}ayesian Inference with Interacting Particle Systems},
+  title={\pkg{Biips}: Software for {B}ayesian {I}nference with {I}nteracting {P}article {S}ystems},
   author={Todeschini, Adrien and Caron, Fran{\c{c}}ois and Fuentes, Marc and Legrand, Pierrick and Del Moral, Pierre},
   journal={arXiv preprint arXiv:1412.3779},
   year={2014}
@@ -38,7 +38,7 @@ journal={arXiv preprint arXiv:1306.3277},
 
 @article{Johansen:2009:JSSOBK:v30i06,
   author =	"Adam M. Johansen",
-  title =	"SMCTC: Sequential Monte Carlo in \proglang{C++}",
+  title =	"SMCTC: {Sequential} {Monte Carlo} in \proglang{C++}",
   journal =	"Journal of Statistical Software",
   volume =	"30",
   number =	"6",
@@ -52,7 +52,7 @@ journal={arXiv preprint arXiv:1306.3277},
 
 @article{doi:10.1080/10618600.2016.1172487,
 author = {Perry de Valpine and Daniel Turek and Christopher J. Paciorek and Clifford Anderson-Bergman and Duncan Temple Lang and Rastislav Bodik},
-title = {Programming With Models: Writing Statistical Algorithms for General Model Structures With \pkg{NIMBLE} },
+title = {{Programming With Models: Writing Statistical Algorithms} for {General Model Structures With} \pkg{NIMBLE} },
 journal = {Journal of Computational and Graphical Statistics},
 volume = {26},
 number = {2},
@@ -99,7 +99,7 @@ doi = {10.1007/978-1-4684-9393-1}
 
 @article {pmcmc,
 author = {Andrieu, Christophe and Doucet, Arnaud and Holenstein, Roman},
-title = {Particle Markov Chain Monte Carlo Methods},
+title = {Particle {Markov} chain {Monte Carlo} methods},
 journal = {Journal of the Royal Statistical Society: Series B (Statistical Methodology)},
 volume = {72},
 number = {3},
@@ -186,7 +186,7 @@ archivePrefix = "arXiv",
 }
 
 @inbook{taylor82,
-title = "Financial returns modelled by the product of two stochastic processes, a study of daily sugar prices",
+title = "Financial {Returns} {Modelled} by the {Product} of {Two} {Stochastic} {Processes}, {A Study of Daily Sugar Prices}",
 author = "Taylor, {Stephen John}",
 note = "Reprinted, original publication 1982, in Time Series Analysis: Theory and Practice 1",
 year = "2018",
@@ -353,14 +353,14 @@ author = {Tran, Minh-Ngoc and Scharth, Marcel and Pitt, Michael and Kohn, Robert
 year = {2013},
 month = {09},
 pages = {},
-title = {Importance Sampling Squared for Bayesian Inference in Latent Variable Models},
+title = {Importance {Sampling Squared} for {Bayesian Inference} in {Latent Variable Models}},
 journal = {SSRN Electronic Journal},
[…]
