### 2020-10-08 #2599 @tbrown122387
@kthyng okey dokey capitalization now matches titles.
