### 2021-02-26 #2599 @tbrown122387
@diehlpk @kthyng is there any chance we can re-publish this after I insert the following blurb somewhere?

> The authors acknowledge Research Computing at The University of Virginia for providing computational resources and technical support that have contributed to the results reported within this publication. URL: https://rc.virginia.edu

These folks had helped me out with some of the `cmake` issues that were raised [here](https://github.com/tbrown122387/pf/issues/13) which was in response to some of the feedback given by @ziotom78 Sorry to make more work for you all. I had just stumbled upon a [page](https://www.rc.virginia.edu/service/acknowledgement/) on their site that requested attribution as a form of support.
