### 2021-02-27 #2599 @arfon
Sure thing! I've added this in https://github.com/openjournals/joss-papers/blob/d5a41afeb6cfbf928dfb18738fd6e74ab1768151/joss.02599/10.21105.joss.02599.pdf
